Re-implement Model class + TermPool + update Term
New Model class is more functional now (rather than state-driven). This
removes the need for a SingletonModel class to keep track of the model
and the need to throw away & replace the model each time parameters are
updated.

TermPool keeps track of created Term instances to avoid duplicates. The
Term caches its column, so this saves time when a term's column is
requested more than once.

Term is updated to compute its column on creation, but cache the result
for future requests. The cache can be invalidated in case of changes.