Change column representation and ordering of terms
Model v2 represents columns by their index in the original data matrix
as opposed to the index in X (the new one includes the dependent column
so that it can be used as an I.V. with lag, and for simplicity). Update
the interface to reflect this.

Order terms by highest t-statistic -> lowest.