Add ability for model to use subsets of the data
Subsets allow for the data to be partitioned for fitting, testing, and
validation. The Model will instantiate with 1 default subset: 'fit',
which will be set to span the entire dataset whenever the data is
updated via `setData()`.

The computed properties of Model and Term have been replaced by methods
which take a subset name as their argument to determine which subset to
operate on.

Likewise, cache properties have been changed to objects so that they can
store results from various subsets.