Move statistical stuff into its own subdirectory
Instead of hard-coding statistical calculations in the `lstsq()`
function, a statistic declares its dependencies (items necessary in
order to compute), and then the statistics are topologically sorted to
be computed in order to ensure dependency satisfaction.

There is some amount of overhead in doing this as opposed to doing it
inline within the `lstsq()` function, but this will allow for
characteristic descriptions of each statistic for the UI in addition to
functional definitions.