Fix interface bugs + sort candidates
When the interface loads, it tells the engine which terms are stored in
localStorage (TODO: need to stall computation until after all terms are
sent).

Some crude sorting is done on the last statistic being shown on the
candidates (the one nearest settings) (TODO: specify which statistic to
sort on).

Naming mismatches fixed.