Use the unsubscribe-from-changes feature on initial load
Surround the block of code that loads everything from localStorage and
passes it to the model with `unsubscribeToChanges()` and
`subscribeToChanges()` to effectively dismiss model and candidate
update messages.

Also, prepare the graph data in a better format and return the
formatting of numbers on the "result" pane back to `num` (as opposed to
`text`, which was set to debug something earlier).