Add "other settings", fix progress bar, 't' & 'p(t)' in global
The only setting to go in other for now is the "subscribe to changes"
toggle, which when switched off holds off on computation until it is
switched back on.

Progress bar was doing `0.xx%` instead of `xx%`. Also, the bezier
timing function is too slow so it should be replaced with `ease`.

`t` and `p(t)` should be prominently displayed next to each term in the
global model so that users know which terms explain how much variance.