Add some observable nonsense
I don't really think this is the proper way to do this, but the model
is now observable, and `get_candidates(Observer)` takes an
Observer<Progress>, which it can update whenever a candidate is
computed. The TF module instance inherits from Observer<Progress> and
Observer<Model>, to get updates for both.

This will likely be turned into a more suitable pattern at some point.