Make caching an extendable mixin, separate lstsqWithStats...
...into lstsq() and statistics(). This allows for more control over the
input to the computation of statistics (e.g. if cross-validation data
should be used for candidates, the estimators from the fit data need to
be passed in with the X and y from the cross data).

`CacheMixin` creates a class that wraps some base class, which is
extendable. This works much better than keeping track of a cache in
each individual class. Instead, the caching mixin handles returning
values for repeated calls with the same set of arguments (and even
default arguments).