Remove usage of webpack-worker plugin
This was making the build slow. Due to an error with webpack-dev-server
calling `postMessage()` on the global scope of a worker and expecting it
to be WebSocket.postMessage (it is Worker.postMessage),
`self.postMessage()` is replaced with a try-catch wrapper that prints
failed calls to `postMessage()` as a warning instead of an error.